+++ app/(navigatable)/(orgs)/orgs/[id]/page.tsx
@@ -14,6 +14,7 @@ import { useSession } from 'next-auth/react'
 
 import { type Org } from './../../../../../lib/data-model-types'
 import { useAppContext } from './../../../../../lib/hooks/app-context'
+import { PremiumPlanUIDescription } from 'components/product-descriptions'
 
 const OrgIndex = ({ params: { id } }: { params: { id: string } }) => {
   const { activeBillingOrg, setActiveBillingOrg } = useAppContext()
@@ -143,26 +144,7 @@ const OrgIndex = ({ params: { id } }: { params: { id: string } }) => {
                 : 'bg-background shadow-md rounded-lg p-6 border mb-4'
             }
           >
-            <div className="flex flex-col items-start">
-              <div className="w-full flex items-center justify-between pb-2">
-                <div className="flex items-center">
-                  <p>Premium Plan</p>
-                  <div
-                    title="Premium Plan"
-                    className="flex items-center justify-center mx-2 p-1 border border-yellow-500 rounded-full"
-                  >
-                    <StarFilledIcon className="w-2 h-2 text-yellow-500" />
-                  </div>
-                </div>
-                <p>100$/Month (per user)</p>
-              </div>
-              <p>✅ Project creation to analyze GitHub repositories</p>
-              <p>✅ Project Goals can be set to guide Sara analysis</p>
-              <p>✅ Sara generated Task-plans to achieve Goals</p>
-              <p>✅ Manual GitHub source synchronization</p>
-              <p>✅ Unlimited project creation</p>
-              <p>✅ Access to private repositories for projects</p>
-            </div>
+            <PremiumPlanUIDescription />
           </div>
         </div>
         {orgIsPremium ? (